after testing execution under root mode: 1. added  bug fixes for parser_implement[] class  methods - add/delete_app_policy(), return_modified_app_policies() method,
the fixes above fix the case when no cap entry for an app exist in the db.

2. fixes in enforcer_implement in buildEnforcerCMDParams() method

the fix enables the construction of the cmd param for a case when no more capabilities remain for an app - therefore we have to remove the last cap from an app living it wi
th no caps. previously this case has not been added due to absence of real testing under root.

3. added drop table statement in the db unit tests to let new tinyPM launch have a blank db instead of having the /bin/ping entry there.