TSL.Utilities:
fragmentEnhanced_ObjectReplica() - working on new method that does not 
depend on the sleep condition and therefore improves replication speed

PTS:
appendEnhanced_ContentTuple() - implemented a new version of append_ContentTuple()
that initially completes the write of the content tuple to hidden file 
and only after that exposes a tuple to public via rename operation.
Such an enhancement eliminates the problem of determining whether the
tuple has been ready for consumption or not by the agent and therefore
avoids potential file corruption during reads. 
This makes append operation safe for sleep-less replica fragmentation by
the controller.