TSL: PersistentTupleSpace_implement:
appendEnhanced_ControlTuple() - implemented the new method and replaced 
as main append_ControlTuple() method and archived the original one.

The enhancement initially completes the write of the control tuple to 
hidden file and only after that exposes a tuple to public via rename 
operation.
Such an enhancement eliminates the problem of determining whether the
tuple has been ready for consumption or not by the agent and therefore
avoids potential file corruption during reads -- race conditions.